pyproject.toml
[tool.poetry]
name = "spa_proof_of_concept"
version = "1.0"
description = "spa-proof-of-concept"
license = "Proprietary" # or "MIT"
authors = ["TODO.com <info@TODO.com>"]
repository = ""

# NOTE:
# without this explicit reference to PyPI, poetry may raise exception "PackageNotFound".
[[tool.poetry.source]]
name = "the_pypi"
url = "https://pypi.org/simple"
default = true

[tool.poetry.dependencies]
python = "^3.9"

Django = "==4.0.*"
environs = "*"
dj-database-url = "*"
django-redis = "==5.2.*"
boto3 = "==1.*"
django-storages = { version = "==1.*" }
gunicorn = "==20.*"
psycopg2-binary = "==2.9.*"
sentry-sdk = "==1.*"
redis = "==4.*"
djangorestframework = "==3.13.*"
djangorestframework_simplejwt = "==5.1.*"
django-cors-headers = "==3.11.*"
django-filter = "==21.*"
tg-react = "==3.*"
tg-utils = {version = "*", extras = ["model-hash", "health_check"]}

[tool.poetry.dev-dependencies]
black = "==22.3.0"
django-debug-toolbar = "*"
# Testing & quality tools
coverage = "==6.*"
ipdb = "*"
safety = "*"
isort = "==5.*"
# Note prospector is very picky about its dependencies,
# avoid specifying versions for pylint/pylint-django/astroid
# and let prospector determine which versions work together
prospector = {version = "==1.*", extras=["with_bandit"]}
pytest = "==7.*"
pytest-django = "*"
pytest-cov = "*"
pytest-mock = "*"
pylint-django = "*"
model-bakery = "*"
freezegun = "*"
# MyPy dependencies
types-freezegun = "*"
boto3-stubs = "*"
django-stubs = "*"
djangorestframework-stubs = "*"
mypy = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
# Enable specific Python version if necessary:
# target-version = ['py38']
line-length = 88
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | docs
    | cover
  )/
  | local.py
  | local_test.py
)
'''

[tool.isort]
skip = "settings/local.py,settings/local_test.py,spa_proof_of_concept/wsgi.py"
skip_glob = "node_modules"
line_length = '88'
atomic = 'true'
multi_line_output = '3'
include_trailing_comma = 'true'
force_grid_wrap = '0'
use_parentheses = 'true'
ensure_newline_before_comments = 'true'
known_django = "django"
known_pytest = "pytest"
overwrite_in_place = 'true'
known_first_party = "accounts,spa_proof_of_concept,settings"
lines_between_types = '0'
sections = "FUTURE,STDLIB,DJANGO,PYTEST,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
lines_after_imports = '2'
force_alphabetical_sort_within_sections = '1'

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    "*test*",
    "*migrations*",
    "manage.py",
    "spa_proof_of_concept/urls.py",
    "spa_proof_of_concept/celery.py",
    "spa_proof_of_concept/celery_settings.py",
    "spa_proof_of_concept/storages.py",
    "spa_proof_of_concept/wsgi.py",
    "settings/*",
]

[tool.coverage.html]
directory = "cover"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--junit-xml report.xml"
DJANGO_SETTINGS_MODULE = "settings.local_test"
markers = ["django_db", "parametrize", "unit"]
norecursedirs = [
    "migrations",
    ".data",
    "app",
    "node_modules",
]
python_files = ["test_*.py", "tests/*.py", "tests.py"]
